requirements.txt
Django==5.2.1
django-environ==0.12.0
pswrd_entropy_gen==1.0.2
psycopg2==2.9.10
djangorestframework==3.16.0
drf-spectacular==0.28.0